requirements.txt
Django==1.11.7
paho-mqtt==1.3.1
graphene-django==2.0.0
django-graphql-jwt==0.0.2

channels==1.1.8
asgi-redis==1.4.3
